Make ColumnarBatchScanExec abstract to break the cyclic dependency
