+++ native-sql-engine/core/src/main/scala/com/intel/oap/extension/ColumnarOverrides.scala
@@ -19,14 +19,15 @@ package com.intel.oap.extension
 
 import com.intel.oap.GazellePluginConfig
 import com.intel.oap.GazelleSparkExtensionsInjector
-import com.intel.oap.sql.shims.SparkShimLoader
 import com.intel.oap.execution._
 import com.intel.oap.extension.columnar.ColumnarGuardRule
 import com.intel.oap.extension.columnar.RowGuard
 import com.intel.oap.sql.execution.RowToArrowColumnarExec
+import com.intel.oap.sql.shims.SparkShimLoader
 
 import org.apache.spark.{MapOutputStatistics, SparkContext}
 import org.apache.spark.internal.Logging
+import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{SparkSession, SparkSessionExtensions}
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.optimizer.BuildLeft
@@ -50,6 +51,8 @@ import org.apache.spark.sql.execution.joins._
 import org.apache.spark.sql.execution.python.{ArrowEvalPythonExec, ColumnarArrowEvalPythonExec}
 import org.apache.spark.sql.execution.window.WindowExec
 import org.apache.spark.sql.internal.SQLConf
+import org.apache.spark.sql.vectorized.ColumnarBatch
+
 import org.apache.spark.util.ShufflePartitionUtils
 
 import scala.collection.mutable
@@ -88,9 +91,27 @@ case class ColumnarPreOverrides() extends Rule[SparkPlan] {
       ColumnarArrowEvalPythonExec(plan.udfs, plan.resultAttrs, columnarChild, plan.evalType)
     case plan: BatchScanExec =>
       logDebug(s"Columnar Processing for ${plan.getClass} is currently supported.")
-      SparkShimLoader.getSparkShims
-        .newColumnarBatchScanExec(plan)
-        .asInstanceOf[ColumnarBatchScanExec]
+      val runtimeFilters = SparkShimLoader.getSparkShims.getRuntimeFilters
+      new ColumnarBatchScanExec(plan.output, plan.scan, runtimeFilters) {
+        // This method is a commonly shared implementation for ColumnarBatchScanExec.
+        // We move it outside of shim layer to break the cyclic dependency caused by
+        // ColumnarDataSourceRDD.
+        override def doExecuteColumnar(): RDD[ColumnarBatch] = {
+          val numOutputRows = longMetric("numOutputRows")
+          val numInputBatches = longMetric("numInputBatches")
+          val numOutputBatches = longMetric("numOutputBatches")
+          val scanTime = longMetric("scanTime")
+          val inputSize = longMetric("inputSize")
+          val inputColumnarRDD =
+            new ColumnarDataSourceRDD(sparkContext, partitions, readerFactory,
+              true, scanTime, numInputBatches, inputSize, tmpDir)
+          inputColumnarRDD.map { r =>
+            numOutputRows += r.numRows()
+            numOutputBatches += 1
+            r
+          }
+        }
+      }
     case plan: CoalesceExec =>
       ColumnarCoalesceExec(plan.numPartitions, replaceWithColumnarPlan(plan.child))
     case plan: InMemoryTableScanExec =>
